+++ tests/test_cofe/test_clusterspace.py
@@ -8,7 +8,7 @@ import json
 from pymatgen.core import Lattice, Structure, Species
 from pymatgen.util.coord import is_coord_subset_pbc
 from pymatgen.symmetry.analyzer import SpacegroupAnalyzer
-from smol.cofe import ClusterSubspace
+from smol.cofe import ClusterSubspace, PottsSubspace
 from smol.cofe.space.clusterspace import invert_mapping,\
                                          get_complete_mapping
 from smol.cofe.extern import EwaldTerm
@@ -17,11 +17,10 @@ from smol.cofe.space.domain import get_allowed_species, Vacancy
 from smol.exceptions import StructureMatchError
 from tests.utils import gen_random_structure
 from src.mc_utils import corr_from_occupancy
+from tests.utils import assert_msonable
 
 
 # TODO test correlations for ternary and for applications of symops to structure
-
-
 def test_invert_mapping_table():
     forward = [[], [], [1], [1], [1], [2, 4], [3, 4], [2, 3], [5, 6, 7]]
     backward = [[], [2, 3, 4], [5, 7], [6, 7], [5, 6], [8], [8], [8], []]
@@ -62,6 +61,36 @@ def test_from_cutoffs(structure):
         npt.assert_allclose(
             np.array(list(subspace.cutoffs.values())),
             np.array(list(tight_subspace.cutoffs.values())))
+
+
+def test_potts_subspace(cluster_subspace):
+    potts_subspace = PottsSubspace.from_cutoffs(cluster_subspace.structure,
+                                                cluster_subspace.cutoffs)
+    assert len(potts_subspace.orbits) == len(cluster_subspace.orbits)
+
+    # check sizes and bits included in each orbit
+    for porbit, corbit in zip(potts_subspace.orbits, cluster_subspace.orbits):
+        assert len(porbit.site_spaces) == len(corbit.site_spaces)
+        assert len(porbit.site_bases) == len(corbit.site_bases)
+        assert len(porbit.bit_combos) > len(corbit.bit_combos)
+
+        for i, site_space in enumerate(porbit.site_spaces):
+            bits_i = np.concatenate([b[:, i] for b in porbit.bit_combos])
+            assert all(j in bits_i for j in site_space.codes)
+
+    # check decorations
+    for _ in range(10):
+        i = random.choice(range(1, potts_subspace.num_corr_functions))
+        o_id = potts_subspace.function_orbit_ids[i]
+        orbit = potts_subspace.orbits[o_id - 1]
+        fdeco = potts_subspace.get_function_decoration(i)
+        odeco = potts_subspace.get_orbit_decorations(o_id)
+        assert fdeco == odeco[i - orbit.bit_id]
+        assert all(  # all decorations include valid species
+            deco[i] in species for deco in fdeco
+            for i, species in enumerate(orbit.site_spaces))
+
+    assert_msonable(potts_subspace)
 
 
 class TestClusterSubSpace(unittest.TestCase):
@@ -371,13 +400,15 @@ class TestClusterSubSpace(unittest.TestCase):
         supercell_struct.make_supercell(m)
         fcoords = np.array(supercell_struct.frac_coords)
 
-        for orb, inds in self.cs.supercell_orbit_mappings(m):
+        for orb, inds in zip(
+                self.cs.orbits, self.cs.supercell_orbit_mappings(m)):
             for x in inds:
                 pbc_radius = np.max(supercell_struct.lattice.get_all_distances(
                     fcoords[x], fcoords[x]))
                 # primitive cell fractional coordinates
                 new_fc = np.dot(fcoords[x], m)
-                self.assertGreater(orb.base_cluster.diameter + 1e-7, pbc_radius)
+                self.assertGreater(
+                    orb.base_cluster.diameter + 1e-7, pbc_radius)
                 found = False
                 for equiv in orb.clusters:
                     if is_coord_subset_pbc(equiv.sites, new_fc, atol=SITE_TOL):
@@ -459,7 +490,7 @@ class TestClusterSubSpace(unittest.TestCase):
         orbit_list = [
             (orb.bit_id, orb.bit_combo_array, orb.bit_combo_inds,
              orb.bases_array, inds)
-            for orb, inds in cs.supercell_orbit_mappings(m)]
+            for orb, inds in zip(cs.orbits, cs.supercell_orbit_mappings(m))]
 
         # last two clusters are switched from CASM output (occupancy basis)
         # all_li (ignore casm point term)
@@ -517,7 +548,7 @@ class TestClusterSubSpace(unittest.TestCase):
         orbit_list = [
             (orb.bit_id, orb.bit_combo_array, orb.bit_combo_inds,
              orb.bases_array, inds)
-            for orb, inds in cs.supercell_orbit_mappings(m)]
+            for orb, inds in zip(cs.orbits, cs.supercell_orbit_mappings(m))]
         # last two pair terms are switched from CASM output (occupancy basis)
         # all_vacancy (ignore casm point term)
         occu = self._encode_occu([Vacancy(),
@@ -569,7 +600,7 @@ class TestClusterSubSpace(unittest.TestCase):
         orbit_list = [
             (orb.bit_id, orb.bit_combo_array, orb.bit_combo_inds,
              orb.bases_array, inds)
-            for orb, inds in cs.supercell_orbit_mappings(m)]
+            for orb, inds in zip(cs.orbits, cs.supercell_orbit_mappings(m))]
         # mixed
         occu = self._encode_occu([Vacancy(), Species('Li', 1),
                                   Species('Li', 1)], self.domains)
@@ -666,12 +697,9 @@ class TestClusterSubSpace(unittest.TestCase):
         # checked that the cached orbit index mappings where properly kept
         for scm, orb_inds in cs._supercell_orb_inds.items():
             self.assertTrue(scm in self.cs._supercell_orb_inds)
-            for orb_inds1, orb_inds2 in zip(orb_inds,
-                                            self.cs._supercell_orb_inds[scm]):
-                self.assertEqual(orb_inds1[0].id, orb_inds1[0].id)
-                self.assertEqual(orb_inds1[0].base_cluster.diameter,
-                                 orb_inds1[0].base_cluster.diameter)
-                self.assertTrue(np.array_equal(orb_inds1[1], orb_inds2[1]))
+            for orb_inds1, orb_inds2 in zip(
+                    orb_inds, self.cs._supercell_orb_inds[scm]):
+                self.assertTrue(np.array_equal(orb_inds1, orb_inds2))
         self.assertTrue(np.array_equal(self.cs.corr_from_structure(struct2),
                                        cs.corr_from_structure(struct2)))
         # Check orthonormalization is kept